Rename service  from mountmanager to mounts
