+++ api/Protobuild.toml
@@ -36,7 +36,7 @@ generators = ["go", "go-ttrpc"]
 [[overrides]]
 prefixes = [
   "github.com/containerd/containerd/api/runtime/sandbox/v1",
-  "github.com/containerd/containerd/api/services/mountmanager/v1",
+  "github.com/containerd/containerd/api/services/mounts/v1",
 ]
 generators = ["go", "go-ttrpc", "go-grpc"]
