+++ services/services.go
@@ -35,6 +35,6 @@ const (
 	IntrospectionService = "introspection-service"
 	// Streaming service is the id of the streaming service
 	StreamingService = "streaming-service"
-	// MountManagerService is the id of the mountmanager service
-	MountManagerService = "mountmanager-service"
+	// MountsService is the id of the mount manager service
+	MountsService = "mounts-service"
 )
